+++ src/GitVersionExe/GitVersionExe.csproj
@@ -64,6 +64,8 @@
     <Compile Include="Arguments.cs" />
     <Compile Include="ExtensionMethods.cs" />
     <Compile Include="HelpWriter.cs" />
+    <Compile Include="Runners\InitRunner.cs" />
+    <Compile Include="Runners\InspectRunner.cs" />
     <Compile Include="Options\InjectBuildServerOptions.cs" />
     <Compile Include="Options\InitOptions.cs" />
     <Compile Include="Options\InspectRemoteRepositoryOptions.cs" />
@@ -75,7 +77,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="AssemblyInfoFileUpdate.cs" />
-    <Compile Include="Runner.cs" />
+    <Compile Include="Runners\InjectBuildServerRunner.cs" />
     <Compile Include="SpecifiedArgumentRunner.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -113,6 +115,7 @@
       <Name>GitVersionCore</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="Clean">
     <ItemGroup>
